requirements.txt
pymupdf 
langchain-huggingface 
faiss-cpu
ollama 
sentence-transformers
sentencepiece
langchain-community


